introduce constants and own class for ExtendedIterator<T> to Stream-Wrapper
