+++ test/de/aksw/package.scala
@@ -8,8 +8,6 @@ import java.net.InetAddress
  */
 package object aksw {
 
-  val dbpediaHostname: String = "dbpedia.org"
-  val DBPEDIA_ENDPOINT: String = s"http://${dbpediaHostname}/sparql"
   val SWAT_RESOURCE_URI: String = "http://dbpedia.org/resource/SWAT"
   val SIMPLESURFACEFORM_TURTLE_MODEL_FILENAME : String = "test-surface-forms.ttl"
